refactor(select-music): rename sort_menu to select_music_menu and split into submodules
Rename the sort_menu module to select_music_menu to better reflect its
scope, and split the 2100-line monolith into focused submodules:

- classic.rs    — original menu rendering (RenderParams, build_overlay)
- categories.rs — stub types for future collapsible-category UI
- song_search.rs — song search overlay (text entry, results, filtering)
- leaderboard.rs — GrooveStats leaderboard overlay
- downloads.rs   — download manager overlay
- replay.rs      — replay selector overlay

Add a gated use_category_select_music_menu config flag (default false)
with State::Classic / State::Categories routing so the new UI can be
developed alongside the old one. Consolidate duplicated action dispatch
into a shared dispatch_menu_action() function.

No user-visible changes — the classic path is active by default and
behavior is identical to before the rename.
