+++ src/config/store/save.rs
@@ -235,6 +235,11 @@ fn push_saved_options(
         "SelectMusicPreviewLoop",
         cfg.select_music_preview_loop,
     );
+    push_bool(
+        content,
+        "UseCategorySelectMusicMenu",
+        cfg.use_category_select_music_menu,
+    );
     push_line(
         content,
         "SelectMusicPatternInfo",
