+++ src/screens/components/select_music/mod.rs
@@ -1,6 +1,6 @@
 pub mod lobby_overlay;
 pub mod music_wheel;
 pub mod screen_bars;
+pub mod select_music_menu;
 pub mod select_pane;
-pub mod sort_menu;
 pub mod step_artist_bar;
